src/utils/common.py
from datetime import datetime
from dateutil import parser

def format_date(date_input: any) -> str:
    """
    Formats a date input into 'YYYY/MM/DD HH:MM:SS' string format.
    It can handle various date string formats, datetime objects, or timestamps.
    """
    if date_input is None:
        return ""

    if isinstance(date_input, str):
        try:
            # Try to parse common timestamp format first if it's a string number
            if date_input.isdigit():
                # Assuming it's a Unix timestamp (seconds)
                dt_object = datetime.fromtimestamp(int(date_input))
            else:
                dt_object = parser.parse(date_input)
        except (ValueError, TypeError, OverflowError):
            return "" # Or raise an error, or return a default
    elif isinstance(date_input, (int, float)): # Assuming it's a Unix timestamp
        try:
            dt_object = datetime.fromtimestamp(date_input)
        except (ValueError, TypeError, OverflowError):
            return "" # Or raise an error, or return a default
    elif isinstance(date_input, datetime):
        dt_object = date_input
    else:
        return "" # Or raise an error for unsupported types

    return dt_object.strftime("%Y/%m/%d %H:%M:%S")

if __name__ == '__main__':
    # Test cases from the original TypeScript
    print(f"'2023-01-01T12:34:56Z': {format_date('2023-01-01T12:34:56Z')}")
    print(f"datetime.now(): {format_date(datetime.now())}")
    ts = datetime.now().timestamp()
    print(f"Timestamp {ts}: {format_date(ts)}")
    print(f"'2023/07/15 10:30:00': {format_date('2023/07/15 10:30:00')}")
    print(f"'Jan 1, 2023 14:45:00': {format_date('Jan 1, 2023 14:45:00')}")
    print(f"None: {format_date(None)}")
    print(f"Empty string: {format_date('')}")
    print(f"Invalid string: {format_date('invalid date')}")
    print(f"1672531200 (timestamp for 2023-01-01T00:00:00Z): {format_date(1672531200)}")
    # Example from original: formatDate(new Date("2023-10-07T18:09:05.969Z"))
    # Example from original: formatDate("2023-10-07T18:09:05.969Z")
    print(f"'2023-10-07T18:09:05.969Z': {format_date('2023-10-07T18:09:05.969Z')}")
    # Example from original: formatDate(new Date("2023-10-07"))
    print(f"'2023-10-07': {format_date('2023-10-07')}")
    # Example from original: formatDate(new Date().getTime())
    print(f"Current timestamp (new Date().getTime()): {format_date(datetime.now().timestamp())}")
